feat: graphql setup complete, add wrapper and util component
