+++ src/app/layout.jsx
@@ -1,3 +1,4 @@
+import ApolloProviderClient from '@/providers/apollo-provider';
 import './globals.css';
 import { JetBrains_Mono } from 'next/font/google';
 
@@ -13,7 +14,11 @@ export const metadata = {
 export default function RootLayout({ children }) {
   return (
     <html suppressHydrationWarning={true} lang='en' className='scroll-smooth'>
-      <body className={jetBrainsMono.className}>{children}</body>
+      <body
+        className={`flex min-h-screen w-full flex-col justify-between bg-main text-white ${jetBrainsMono.className}`}
+      >
+        <ApolloProviderClient>{children}</ApolloProviderClient>
+      </body>
     </html>
   );
 }
