--- src/data/sample-data.json
@@ -1,50 +1,0 @@
-[
-  {
-    "date": "Sun Jan 02 2022 00:00:00 GMT+0000 (Greenwich Mean Time)",
-    "total": 302
-  },
-  {
-    "date": "Sun Jan 09 2022 00:00:00 GMT+0000 (Greenwich Mean Time)",
-    "total": 196
-  },
-  {
-    "date": "Sun Jan 16 2022 00:00:00 GMT+0000 (Greenwich Mean Time)",
-    "total": 143
-  },
-  {
-    "date": "Sun Jan 23 2022 00:00:00 GMT+0000 (Greenwich Mean Time)",
-    "total": 102
-  },
-  {
-    "date": "Sun Jan 30 2022 00:00:00 GMT+0000 (Greenwich Mean Time)",
-    "total": 20
-  },
-  {
-    "date": "Sun Feb 06 2022 00:00:00 GMT+0000 (Greenwich Mean Time)",
-    "total": 42
-  },
-  {
-    "date": "Sun Feb 13 2022 00:00:00 GMT+0000 (Greenwich Mean Time)",
-    "total": 2
-  },
-  {
-    "date": "Sun Feb 20 2022 00:00:00 GMT+0000 (Greenwich Mean Time)",
-    "total": 122
-  },
-  {
-    "date": "Sun Feb 27 2022 00:00:00 GMT+0000 (Greenwich Mean Time)",
-    "total": 203
-  },
-  {
-    "date": "Sun Mar 06 2022 00:00:00 GMT+0000 (Greenwich Mean Time)",
-    "total": 202
-  },
-  {
-    "date": "Sun Mar 13 2022 00:00:00 GMT+0000 (Greenwich Mean Time)",
-    "total": 20
-  },
-  {
-    "date": "Sun Mar 20 2022 00:00:00 GMT+0000 (Greenwich Mean Time)",
-    "total": 22
-  }
-]
