+++ src/providers/apollo-provider.jsx
@@ -1,0 +1,15 @@
+'use client';
+
+import { ApolloClient, ApolloProvider, InMemoryCache } from '@apollo/client';
+
+const client = new ApolloClient({
+  uri: 'https://api.github.com/graphql',
+  cache: new InMemoryCache(),
+  headers: {
+    authorization: `Bearer ${process.env.NEXT_PUBLIC_PERSONAL_ACCESS_TOKEN}`,
+  },
+});
+
+export default function ApolloProviderClient({ children }) {
+  return <ApolloProvider client={client}>{children}</ApolloProvider>;
+}
